bug fixes to fostercar code
The first bug was in the cardEffectAdventurer method.  Specifically at line 1197 of dominion.c.  This while loop represents discarding the drawn cards.  Since that means the card should reduce, the value of z should be z = z – 1.  The error is z = z + 1.

Second bug in cardEffectSmithy.  Specifically at line 1211 of dominion.c.  The discardCard  should  pass in a value of 0 in the 4th paramenter, but the error  in the code is value of 1